Write files only on transaction commit
Otherwise if an error occurs e.g. inside the MessageStored event
handler, the amqp transaction is aborted (so the same message will be
processed again some time later), but the file has already been
written. This means another file with the same contents will be
written when the message is processed again (since usually filenames
are generated to include a uniquifiying element).