Introduce ``IHandler.exception()`` for transaction-integrated error handling
If an error occurs during the transaction.commit, previously it was
impossible to perform error handling that needs a commit (e.g. send
back an error message, write to a database, etc.), since the
transaction was aborted due to the error, and that was the end of the
story. Now we add another (error-handling) transaction after the main
transaction.